# routes/users.js
const express = require('express');
const router = express.Router();

const supabase = require('../supabase');

const { checkAuthenticated, checkNotAuthenticated } = require('../passport/passportAuth');

router.get('/', checkAuthenticated, (req, res, next) => {
	const user = {
		email: req.user.email,
		language: req.user.language,
		darktheme: req.user.darktheme,
		googleName: req.user.googleName,
		avatarVariant: req.user.avatarVariant,
		avatarColors: req.user.avatarColors
	};
	res.status(200).send(user);
});

router.put('/', checkAuthenticated, async (req, res, next) => {
	const { data, error } = await supabase
		.from('users')
		.update(req.body)
		.eq('uuid', req.user.uuid);

	if (error) {
		errorMessage = new Error('Database failed to update a user.');
		console.log(errorMessage);
		console.error(error);
		res.status(502).send(errorMessage);
		return;
	}

	res.status(200).send(req.body);
});

router.delete('/', checkAuthenticated, async (req, res, next) => {

	if ((req.user.role = 'admin')) {
		errorMessage = 'Cannot delete admin';
		console.error(errorMessage);
		res.status(502).send(errorMessage);
		return;
	}

	const errorUser = await supabase.from('users').delete().eq('uuid', req.user.uuid).error;

	if (errorUser) {
		errorMessage = 'Database delete operation failed';
		console.error(errorMessage);
		console.log(errorUser);
		res.status(502).send(errorMessage);
		return;
	}

	res.status(200).send('User successfully deleted');
});

module.exports = router;